user fields is configured

+++ db/migrate/20140201213907_add_columns_to_users.rb
@@ -1,7 +1,7 @@
 class AddColumnsToUsers < ActiveRecord::Migration
   def change
     add_column :users, :name, :string
-    add_column :users, :phone, :decimal
-    add_column :users, :celphone, :decimal
+    add_column :users, :phone, :integer
+    add_column :users, :celphone, :integer
   end
 end
